requirements.txt
python-telegram-bot
psycopg2-binary
SQLAlchemy
